backend/service-whatsapp-browser/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
playwright>=1.49.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
aiohttp>=3.10.0
motor>=3.6.0
wsproto>=1.2.0
grpcio>=1.66.0
grpcio-reflection>=1.66.0
protobuf>=5.34,<7
